Uni-variate variable selection

Uni-variate variable selection

analyse_univariate(
  x,
  response = NULL,
  vars = NULL,
  level = 0.05,
  family = stats::gaussian(),
  method = "Wald"
)

Arguments

x

An output of the bootstrap_data function.

response

a character argument denoting the outcome

vars

a character vector denoting the covariates or features to be screened

level

a numeric scalar denoting a significance threshold to be used against each uni-variate test performed. Default is 0.05.

family

the error distribution and link function passed to the glm() call inside the analyse_univariate function.

method

a character value of the chosen test to be used. Currently only a Wald chi-squared test is implemented.

Value

A nested tibble

Details

A number of analysis_ wrappers have been created to illustrate the challenge of final variable selection. These can be combined as desired by the user.

Examples

datasets::mtcars %>% 
  bootstrap_data(10) %>% 
  analyse_univariate(response = "mpg",
                     vars = c("disp", "hp","drat", "wt","qsec"),
                     level = 0.05)
#> # A tibble: 10 × 3
#>    boot_rep data           vars     
#>       <int> <list>         <list>   
#>  1        1 <df [32 × 11]> <chr [5]>
#>  2        2 <df [32 × 11]> <chr [5]>
#>  3        3 <df [32 × 11]> <chr [5]>
#>  4        4 <df [32 × 11]> <chr [5]>
#>  5        5 <df [32 × 11]> <chr [5]>
#>  6        6 <df [32 × 11]> <chr [5]>
#>  7        7 <df [32 × 11]> <chr [5]>
#>  8        8 <df [32 × 11]> <chr [4]>
#>  9        9 <df [32 × 11]> <chr [5]>
#> 10       10 <df [32 × 11]> <chr [5]>